py: Factor grid definition into its own file.
